requirements.txt
numpy==1.25.2
pandas==2.1.0
pyarrow==13.0.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
setuptools==58.1.0
six==1.16.0
tzdata==2023.3